Fix task_hook_wrapper
When we run that module against an IOC, the value of deser_data is a list containing an integer and an object. Inside the for loop, the dse_data takes the integer value and cannot be merged inside the database. We need the object from that list. Added check for filtering out the integer values.