[806][FIX] Ensure passwords are aphanumeric
Special characters in passwords may break some functionality. Celery
seems to have an issue with special characters (either `+` or `/`) that
cause the worker to exit without displaying any errors. Attempts to
URL-escape the password seem to fail since the value is interpolated at
some point and Python throws an exception since some `%XX` escape codes
are not valid syntax.

This change reimplements the `validate_config` function to check that
the POSTGRES password / admin password and IRIS secret key / password
salt are long (longer than 30 characters) alphanumeric strings. It
should not affect users of the "scripts/iris_helper.sh" helper script
since the values generated by it are 32 characters long.

The old implementation was removed because the function was not called and
the check `section not in self` looks invalid.

Closes #806

Signed-off-by: Darius Mihai <dariusmihaim@gmail.com>